Default pan now works, pan for harmonized voices now is correct.

+++ src/clarinoid/harmonizer/harmonizer.hpp
@@ -98,7 +98,7 @@ struct Harmonizer
       pout->mIsNoteCurrentlyMuted = false;
       pout->mVoiceId = MakeMusicalVoiceID(loopLayerID, (uint8_t)(nVoice + 1)); // +1 because live voice is id 0.
 
-      pout->mPan += preset.mStereoSeparation * (((nVoice & 1) * 2) - 1); // turns bit 0 to -1 or 1
+      pout->mPan += preset.mStereoSeparation * ((((int)nVoice & 1) * 2) - 1); // turns bit 0 to -1 or 1
 
       switch (hv.mPitchBendParticipation) {
         case PitchBendParticipation::Off:
